+++ cucumber-expressions/javascript/src/ParameterTypeRegistry.ts
@@ -125,5 +125,3 @@ export default class ParameterTypeRegistry {
     }
   }
 }
-
-module.exports = ParameterTypeRegistry