+++ cucumber-expressions/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 * Remove deprecated `CucumberExpressionGenerator#generateExpression` method.
+  ([#1752](https://github.com/cucumber/common/pull/1752))
 
 ### Fixed